requirements.txt
djangorestframework >= 3.15.1
psycopg2-binary >= 2.9.9
psycopg2 >= 2.9.9
pytest >= 8.2.0
Django>=4.0
psycopg2<=3.0




